requirements.txt
blinker==1.9.0
click==8.1.8
exceptiongroup==1.2.2
Flask==3.1.0
flask-cors==5.0.1
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.5
opencv-python==4.11.0.86
packaging==25.0
pillow==11.2.1
pluggy==1.5.0
pytesseract==0.3.13
pytest==8.3.5
tomli==2.2.1
Werkzeug==3.1.3
